Added ArrayList to Order

+++ src/main/java/edu/metrostate/Item.java
@@ -2,11 +2,11 @@ package edu.metrostate;
 
 public class Item {
     private String name;
-    private float price;
+    private double price;
     private int quantity;
 
     //constructor
-    public Item(String name, float price, int quantity) {
+    public Item(String name, double price, int quantity) {
         this.name = name;
         this.price = price;
         this.quantity = quantity;
@@ -14,11 +14,12 @@ public class Item {
 
     //getters
     public String getName() {return name;}
-    public float getPrice() {return price;}
+    public double getPrice() {return price;}
     public int getQuantity() {return quantity;}
 
     //setters
     public void setName(String name) {this.name = name;}
     public void setPrice(float price) {this.price = price;}
     public void setQuantity(int quantity) {this.quantity = quantity;}
+
 }
